requirements.txt
fastapi==0.115.11
uvicorn==0.24.0
pydantic==2.10.6
python-dotenv==1.0.1
structlog==25.2.0
redis==5.2.1
docker==7.1.0
requests==2.32.3
kubernetes==32.0.1